Logicals can't be interpolated, causing error when defining non-erodible layer as logicals. ne_w is now interpolated as double and then further evaluated in case of logical input.
svn path=/trunk/matlab/; revision=18927


Former-commit-id: 1f0611cf4d7cf86573bd3c96709d2b385d68366a

+++ applications/xbeach/xb_modelsetup/xb_grid/xb_grid_optimize.m
@@ -325,7 +325,7 @@ if isvector(z_w)
         
         xb_verbose(1,'Add non-erodible layer');
         
-        negrid = interp1(x_d, ne_w, x_xb);
+        negrid = interp1(x_d, double(ne_w), x_xb);
         if islogical(OPT.ne)
             negrid(isnan(negrid)) = 0;
             idx = ~logical(round(negrid));
@@ -369,7 +369,7 @@ else
         
         xb_verbose(1,'Add non-erodible layer');
         
-        negrid = interp2(x_w, y_w, ne_w, x_xb_w, y_xb_w);
+        negrid = interp2(x_w, y_w, double(ne_w), x_xb_w, y_xb_w);
         if islogical(OPT.ne)
             negrid(isnan(negrid)) = 0;
             idx = ~logical(round(negrid));
